requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pytesseract==0.3.10
pdfplumber==0.10.3
Pillow==10.1.0
PyPDF2==3.0.1
python-multipart==0.0.6